# Make key_prefix::get_shared_length branchless