Use own inline assembly instead of __builtin_ffs in order to remove redundant
CMOVE out of BSF/CMOVE pair. This will be redundant with GCC 11 and/or on
TZCNT-supporting architectures.

Performance of micro_benchmark_key_prefix: baseline:

2020-06-13 09:38:29
Running ./micro_benchmark_key_prefix
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.00, 0.02, 0.02
------------------------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------
unpredictable_get_shared_length      0.949 us        0.949 us       737605 items_per_second=22.1218M/s

 Performance counter stats for './micro_benchmark_key_prefix':

          1,538.31 msec task-clock                #    0.988 CPUs utilized
                 2      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
               150      page-faults               #    0.098 K/sec
     5,237,746,859      cycles                    #    3.405 GHz                      (83.36%)
     2,829,300,464      stalled-cycles-frontend   #   54.02% frontend cycles idle     (83.36%)
     2,090,210,528      stalled-cycles-backend    #   39.91% backend cycles idle      (66.72%)
     3,750,534,408      instructions              #    0.72  insn per cycle
                                                  #    0.75  stalled cycles per insn  (83.36%)
       687,619,152      branches                  #  446.995 M/sec                    (83.36%)
        24,172,616      branch-misses             #    3.52% of all branches          (83.21%)

       1.557020749 seconds time elapsed

       1.057789000 seconds user
       0.480813000 seconds sys

With the change:

unpredictable_get_shared_length      0.849 us        0.850 us       823508 items_per_second=24.7076M/s

 Performance counter stats for './micro_benchmark_key_prefix':

          1,596.06 msec task-clock                #    0.989 CPUs utilized
                 2      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
               146      page-faults               #    0.091 K/sec
     5,434,776,483      cycles                    #    3.405 GHz                      (83.21%)
     3,055,675,294      stalled-cycles-frontend   #   56.22% frontend cycles idle     (83.21%)
     2,073,259,164      stalled-cycles-backend    #   38.15% backend cycles idle      (66.92%)
     4,194,278,487      instructions              #    0.77  insn per cycle
                                                  #    0.73  stalled cycles per insn  (83.46%)
       689,404,657      branches                  #  431.942 M/sec                    (83.46%)
        16,659,388      branch-misses             #    2.42% of all branches          (83.20%)

       1.614440806 seconds time elapsed

       1.135083000 seconds user
       0.461253000 seconds sys

Cachegrind with branch prediction simulation: baseline:

unpredictable_get_shared_length       14.1 us         14.2 us        49447 items_per_second=1.48367M/s
==833590==
==833590== I   refs:      172,216,844
==833590== I1  misses:          8,228
==833590== LLi misses:          4,039
==833590== I1  miss rate:        0.00%
==833590== LLi miss rate:        0.00%
==833590==
==833590== D   refs:       54,744,658  (38,850,773 rd   + 15,893,885 wr)
==833590== D1  misses:         21,665  (    18,524 rd   +      3,141 wr)
==833590== LLd misses:         11,164  (     9,176 rd   +      1,988 wr)
==833590== D1  miss rate:         0.0% (       0.0%     +        0.0%  )
==833590== LLd miss rate:         0.0% (       0.0%     +        0.0%  )
==833590==
==833590== LL refs:            29,893  (    26,752 rd   +      3,141 wr)
==833590== LL misses:          15,203  (    13,215 rd   +      1,988 wr)
==833590== LL miss rate:          0.0% (       0.0%     +        0.0%  )
==833590==
==833590== Branches:       26,805,998  (26,421,305 cond +    384,693 ind)
==833590== Mispredicts:     1,888,051  ( 1,639,976 cond +    248,075 ind)
==833590== Mispred rate:          7.0% (       6.2%     +       64.5%   )

With the change:

unpredictable_get_shared_length       13.1 us         13.2 us        53058 items_per_second=1.59182M/s
==834387==
==834387== I   refs:      186,766,320
==834387== I1  misses:          7,925
==834387== LLi misses:          3,986
==834387== I1  miss rate:        0.00%
==834387== LLi miss rate:        0.00%
==834387==
==834387== D   refs:       59,994,244  (43,493,204 rd   + 16,501,040 wr)
==834387== D1  misses:         21,664  (    18,525 rd   +      3,139 wr)
==834387== LLd misses:         11,164  (     9,175 rd   +      1,989 wr)
==834387== D1  miss rate:         0.0% (       0.0%     +        0.0%  )
==834387== LLd miss rate:         0.0% (       0.0%     +        0.0%  )
==834387==
==834387== LL refs:            29,589  (    26,450 rd   +      3,139 wr)
==834387== LL misses:          15,150  (    13,161 rd   +      1,989 wr)
==834387== LL miss rate:          0.0% (       0.0%     +        0.0%  )
==834387==
==834387== Branches:       23,195,601  (22,789,254 cond +    406,347 ind)
==834387== Mispredicts:     1,118,359  (   855,970 cond +    262,389 ind)
==834387== Mispred rate:          4.8% (       3.8%     +       64.6%   )

The other benchmarks regress because they all have key prefix lengths of zero,
which get well-predicted through a shorter instruction sequence.

micro_benchmark_node4 baseline:

full_node4_sequential_insert/100         8.89 us         8.91 us        78700 +4=34 16=0 16^=0 256=0 4=34 48=0 48^=0 4^=0 KPfS=9 L=100 items_per_second=11.2293M/s size=12.6289k
full_node4_sequential_insert/512         46.0 us         46.1 us        15209 +4=171 16=0 16^=0 256=0 4=171 48=0 48^=0 4^=0 KPfS=43 L=512 items_per_second=11.1153M/s size=64.5156k
full_node4_sequential_insert/4096         388 us          388 us         1805 +4=1.365k 16=0 16^=0 256=0 4=1.365k 48=0 48^=0 4^=0 KPfS=341 L=4.096k items_per_second=10.5513M/s size=515.984k
full_node4_sequential_insert/32768       4072 us         4072 us          172 +4=10.923k 16=0 16^=0 256=0 4=10.923k 48=0 48^=0 4^=0 KPfS=2.731k L=32.768k items_per_second=8.04753M/s size=4.03127M
full_node4_sequential_insert/65535       9095 us         9095 us           77 +4=21.845k 16=0 16^=0 256=0 4=21.845k 48=0 48^=0 4^=0 KPfS=5.461k L=65.535k items_per_second=7.20574M/s size=8.06238M
full_node4_random_insert/100             11.9 us         11.9 us        58877 items_per_second=8.40695M/s
full_node4_random_insert/512             57.2 us         57.2 us        12204 items_per_second=8.94932M/s
full_node4_random_insert/4096             472 us          472 us         1483 items_per_second=8.67693M/s
full_node4_random_insert/32768           4725 us         4725 us          148 items_per_second=6.93473M/s
full_node4_random_insert/65535          10353 us        10353 us           68 items_per_second=6.33032M/s
node4_full_scan/100                      2.76 us         2.76 us       253656 items_per_second=36.2369M/s
node4_full_scan/512                      17.8 us         17.8 us        39329 items_per_second=28.7648M/s
node4_full_scan/4096                      170 us          170 us         4123 items_per_second=24.1233M/s
node4_full_scan/32768                    1787 us         1787 us          392 items_per_second=18.3356M/s
node4_full_scan/65535                    3783 us         3783 us          185 items_per_second=17.3255M/s
node4_random_gets/100                    57.1 us         57.3 us        12232 items_per_second=17.4612k/s
node4_random_gets/512                     298 us          299 us         2344 items_per_second=3.34803k/s
node4_random_gets/4096                   2467 us         2474 us          283 items_per_second=404.156/s
node4_random_gets/32768                 20843 us        20911 us           33 items_per_second=47.8214/s
node4_random_gets/65535                 41769 us        41874 us           17 items_per_second=23.8814/s
full_node4_sequential_delete/100         6.77 us         6.77 us       103114 items_per_second=14.7792M/s
full_node4_sequential_delete/512         34.9 us         34.9 us        20064 items_per_second=14.6842M/s
full_node4_sequential_delete/4096         319 us          319 us         2198 items_per_second=12.8583M/s
full_node4_sequential_delete/32768       2882 us         2881 us          243 items_per_second=11.372M/s
full_node4_sequential_delete/65535       6055 us         6055 us          116 items_per_second=10.8233M/s
full_node4_random_deletes/100            8.66 us         8.65 us        80909 items_per_second=11.5598M/s
full_node4_random_deletes/512            46.3 us         46.3 us        15138 items_per_second=11.0677M/s
full_node4_random_deletes/4096            439 us          439 us         1596 items_per_second=9.33982M/s
full_node4_random_deletes/32768          5098 us         5098 us          137 items_per_second=6.42746M/s
full_node4_random_deletes/65535         13743 us        13743 us           51 items_per_second=4.76845M/s

 Performance counter stats for './micro_benchmark_node4':

         48,909.02 msec task-clock                #    1.000 CPUs utilized
                48      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
           387,014      page-faults               #    0.008 M/sec
   166,551,872,028      cycles                    #    3.405 GHz                      (83.33%)
    68,129,597,615      stalled-cycles-frontend   #   40.91% frontend cycles idle     (83.33%)
    36,160,081,019      stalled-cycles-backend    #   21.71% backend cycles idle      (66.66%)
   275,159,220,739      instructions              #    1.65  insn per cycle
                                                  #    0.25  stalled cycles per insn  (83.33%)
    57,845,829,581      branches                  # 1182.723 M/sec                    (83.34%)
       396,869,983      branch-misses             #    0.69% of all branches          (83.33%)

      48.927147215 seconds time elapsed

      43.940919000 seconds user
       4.968556000 seconds sys

With the change:

full_node4_sequential_insert/100         8.84 us         8.85 us        78953 +4=34 16=0 16^=0 256=0 4=34 48=0 48^=0 4^=0 KPfS=9 L=100 items_per_second=11.2965M/s size=12.6289k
full_node4_sequential_insert/512         46.4 us         46.4 us        15082 +4=171 16=0 16^=0 256=0 4=171 48=0 48^=0 4^=0 KPfS=43 L=512 items_per_second=11.0243M/s size=64.5156k
full_node4_sequential_insert/4096         406 us          406 us         1724 +4=1.365k 16=0 16^=0 256=0 4=1.365k 48=0 48^=0 4^=0 KPfS=341 L=4.096k items_per_second=10.087M/s size=515.984k
full_node4_sequential_insert/32768       4382 us         4382 us          160 +4=10.923k 16=0 16^=0 256=0 4=10.923k 48=0 48^=0 4^=0 KPfS=2.731k L=32.768k items_per_second=7.47802M/s size=4.03127M
full_node4_sequential_insert/65535       9927 us         9927 us           70 +4=21.845k 16=0 16^=0 256=0 4=21.845k 48=0 48^=0 4^=0 KPfS=5.461k L=65.535k items_per_second=6.60164M/s size=8.06238M
full_node4_random_insert/100             11.9 us         11.9 us        58632 items_per_second=8.37749M/s
full_node4_random_insert/512             59.7 us         59.7 us        11715 items_per_second=8.57103M/s
full_node4_random_insert/4096             507 us          507 us         1379 items_per_second=8.07767M/s
full_node4_random_insert/32768           5276 us         5275 us          134 items_per_second=6.21149M/s
full_node4_random_insert/65535          11561 us        11560 us           61 items_per_second=5.66888M/s
node4_full_scan/100                      3.36 us         3.36 us       207990 items_per_second=29.7201M/s
node4_full_scan/512                      21.8 us         21.8 us        32043 items_per_second=23.4388M/s
node4_full_scan/4096                      209 us          209 us         3342 items_per_second=19.5566M/s
node4_full_scan/32768                    2253 us         2253 us          311 items_per_second=14.547M/s
node4_full_scan/65535                    4706 us         4706 us          149 items_per_second=13.9263M/s
node4_random_gets/100                    57.6 us         57.8 us        12118 items_per_second=17.304k/s
node4_random_gets/512                     302 us          303 us         2313 items_per_second=3.30482k/s
node4_random_gets/4096                   2525 us         2534 us          277 items_per_second=394.707/s
node4_random_gets/32768                 21511 us        21587 us           32 items_per_second=46.3249/s
node4_random_gets/65535                 43009 us        43164 us           16 items_per_second=23.1677/s
full_node4_sequential_delete/100         7.06 us         7.06 us        99047 items_per_second=14.1595M/s
full_node4_sequential_delete/512         37.3 us         37.4 us        18736 items_per_second=13.7064M/s
full_node4_sequential_delete/4096         346 us          346 us         2021 items_per_second=11.8308M/s
full_node4_sequential_delete/32768       3196 us         3196 us          219 items_per_second=10.2518M/s
full_node4_sequential_delete/65535       6794 us         6794 us          103 items_per_second=9.64608M/s
full_node4_random_deletes/100            8.80 us         8.79 us        79554 items_per_second=11.371M/s
full_node4_random_deletes/512            48.2 us         48.2 us        14509 items_per_second=10.6168M/s
full_node4_random_deletes/4096            476 us          476 us         1470 items_per_second=8.60192M/s
full_node4_random_deletes/32768          5650 us         5650 us          124 items_per_second=5.79934M/s
full_node4_random_deletes/65535         14877 us        14876 us           47 items_per_second=4.40529M/s

 Performance counter stats for './micro_benchmark_node4':

         47,954.78 msec task-clock                #    1.000 CPUs utilized
               108      context-switches          #    0.002 K/sec
                 2      cpu-migrations            #    0.000 K/sec
           375,302      page-faults               #    0.008 M/sec
   163,267,007,933      cycles                    #    3.405 GHz                      (83.33%)
    70,633,989,291      stalled-cycles-frontend   #   43.26% frontend cycles idle     (83.33%)
    38,387,456,518      stalled-cycles-backend    #   23.51% backend cycles idle      (66.67%)
   263,929,272,273      instructions              #    1.62  insn per cycle
                                                  #    0.27  stalled cycles per insn  (83.33%)
    50,871,078,965      branches                  # 1060.814 M/sec                    (83.33%)
       330,748,698      branch-misses             #    0.65% of all branches          (83.33%)

      47.973920664 seconds time elapsed

      42.895401000 seconds user
       5.059929000 seconds sys

Baseline ./micro_benchmark:

dense_insert_mem_check/100                                8.25 us         8.27 us        84536 items_per_second=12.0862M/s size=13.0508k
dense_insert_mem_check/512                                36.2 us         36.2 us        19317 items_per_second=14.1365M/s size=60.5781k
dense_insert_mem_check/4096                                278 us          278 us         2516 items_per_second=14.7302M/s size=484.406k
dense_insert_mem_check/32768                              3562 us         3562 us          197 items_per_second=9.20035M/s size=3.78517M
dense_insert_mem_check/262144                            32867 us        32866 us           21 items_per_second=7.97606M/s size=30.2735M
dense_insert_mem_check/2097152                          164837 us       164837 us            4 items_per_second=12.7226M/s size=242.189M
dense_insert_mem_check/16777216                        1994372 us      1994331 us            1 items_per_second=8.41245M/s size=1.8921G
dense_insert_mem_check/30000000                        3068481 us      3068470 us            1 items_per_second=9.77686M/s size=3.38333G
dense_insert_no_mem_check/100                             8.72 us         8.73 us        81145 +4=1 16=0 16^=1 256=1 4=0 48=0 48^=1 4^=1 KPfS=0 L=100 items_per_second=11.4545M/s
dense_insert_no_mem_check/512                             36.3 us         36.3 us        19288 +4=3 16=0 16^=2 256=2 4=1 48=0 48^=2 4^=2 KPfS=1 L=512 items_per_second=14.0981M/s
dense_insert_no_mem_check/4096                             286 us          286 us         2445 +4=17 16=1 16^=16 256=16 4=0 48=0 48^=16 4^=17 KPfS=1 L=4.096k items_per_second=14.3081M/s
dense_insert_no_mem_check/32768                           2254 us         2254 us          311 +4=129 16=0 16^=129 256=129 4=0 48=0 48^=129 4^=129 KPfS=1 L=32.768k items_per_second=14.5369M/s
dense_insert_no_mem_check/262144                         18514 us        18514 us           38 +4=1029 16=0 16^=1028 256=1028 4=1 48=0 48^=1028 4^=1028 KPfS=5 L=262.144k items_per_second=14.1594M/s
dense_insert_no_mem_check/2097152                       148392 us       148391 us            5 +4=8.225k 16=0 16^=8.225k 256=8.224k 4=0 48=1 48^=8.224k 4^=8.225k KPfS=33 L=2.09715M items_per_second=14.1326M/s
dense_insert_no_mem_check/16777216                     1180809 us      1180781 us            1 +4=65.793k 16=0 16^=65.793k 256=65.793k 4=0 48=0 48^=65.793k 4^=65.793k KPfS=257 L=16.7772M items_per_second=14.2086M/s
dense_insert_no_mem_check/30000000                     2612583 us      2612559 us            1 +4=117.649k 16=0 16^=117.648k 256=117.648k 4=1 48=0 48^=117.648k 4^=117.648k KPfS=461 L=30M items_per_second=11.483M/s
sparse_insert_mem_check_dups_allowed/100                  65.2 us         65.5 us        10696 items_per_second=1.52766M/s size=13.8477k
sparse_insert_mem_check_dups_allowed/512                   325 us          327 us         2141 items_per_second=1.56574M/s size=66.8281k
sparse_insert_mem_check_dups_allowed/4096                 2632 us         2645 us          265 items_per_second=1.54865M/s size=546.078k
sparse_insert_mem_check_dups_allowed/32768               20847 us        20942 us           33 items_per_second=1.56468M/s size=4.31255M
sparse_insert_mem_check_dups_allowed/262144             180438 us       181827 us            4 items_per_second=1.44172M/s size=34.1208M
sparse_insert_mem_check_dups_allowed/2097152           1601246 us      1612184 us            1 items_per_second=1.30081M/s size=272.991M
sparse_insert_mem_check_dups_allowed/10000000          7857562 us      7912942 us            1 items_per_second=1.26375M/s size=1.27014G
sparse_insert_no_mem_check_dups_allowed/100               67.6 us         68.0 us        10292 +4=14 16=0 16^=1 256=1 4=13 48=0 48^=1 4^=1 KPfS=0 L=100 items_per_second=1.47084M/s
sparse_insert_no_mem_check_dups_allowed/512                334 us          336 us         2083 +4=151 16=15 16^=1 256=1 4=135 48=0 48^=1 4^=16 KPfS=0 L=512 items_per_second=1.52264M/s
sparse_insert_no_mem_check_dups_allowed/4096              2701 us         2718 us          258 +4=373 16=160 16^=97 256=1 4=116 48=96 48^=1 4^=257 KPfS=1 L=4.096k items_per_second=1.50693M/s
sparse_insert_no_mem_check_dups_allowed/32768            21563 us        21687 us           32 +4=6.202k 16=7 16^=257 256=257 4=5.938k 48=0 48^=257 4^=264 KPfS=3 L=32.768k items_per_second=1.51094M/s
sparse_insert_no_mem_check_dups_allowed/262144          184445 us       186091 us            4 +4=61.752k 16=23.969k 16^=257 256=257 4=37.526k 48=0 48^=257 4^=24.226k KPfS=179 L=262.144k items_per_second=1.40869M/s
…